Core: Contentmanager - Non-admin/ non-owner delete page functionality implemented Non-validated pages removed from 'Ãœbergeordnete Seite' ContentWortkflow - Admin approve/reject functionality for New, update and delete pages corrected
git-svn-id: http://svn.contrexxlabs.com/contrexx/trunk@8271 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ core/ContentManager.class.php
@@ -2421,7 +2421,7 @@ class ContentManager
             }
 
             /*
-             * Fetched username to allow owner delete page
+             * Fetched username to allow owner delete page before admin approval
              */
             $objResult = $objDatabase->Execute("
                 SELECT parcat, catid, module, username
@@ -2432,6 +2432,7 @@ class ContentManager
             if ($objResult !== false && $objResult->RecordCount()>0) {
                 $moduleId = $objResult->fields['module'];
                 $pageOwner = $objResult->fields['username'];
+                $objFWUser = FWUser::getFWUserObject();
                 // needed for recordcount
                 while (!$objResult->EOF) {
                     $objResult->MoveNext();
@@ -2448,16 +2449,32 @@ class ContentManager
                              /*
                              * Delete update logs, Remove revisions from admin approval list
                              */
-                            $objResult = $objDatabase->Execute('
-                                SELECT id
-                                  FROM '.DBPREFIX.'content_navigation_history
-                                 WHERE catid='.$pageId); //.' AND `lang`='.$this->langId
-                            while(!$objResult->EOF) {
-                                $objDatabase->Execute('
-                                    DELETE FROM '.DBPREFIX.'content_logfile
-                                        WHERE   history_id ='.$objResult->fields["id"]);
-                                $objResult->MoveNext();
-                            }
+                             if($pageOwner != $objFWUser->objUser->getUsername() && !$objFWUser->objUser->getAdminStatus()) {
+                                $objResult = $objDatabase->Execute('
+                                    SELECT id
+                                        FROM '.DBPREFIX.'content_navigation_history
+                                        WHERE catid='.$pageId.'
+                                        AND is_active="1"'); //.' AND `lang`='.$this->langId
+                                    while(!$objResult->EOF) {
+                                        $objDatabase->Execute('
+                                            INSERT INTO '.DBPREFIX.'content_logfile
+                                               SET action="delete",
+                                                   history_id='.$objResult->fields['id'].',
+                                                   is_validated="'.($this->boolHistoryActivate ? 1 : 0).'"');
+                                        $objResult->MoveNext();
+                                    }
+                                } else {
+                                    $objResult = $objDatabase->Execute('
+                                        SELECT id
+                                          FROM '.DBPREFIX.'content_navigation_history
+                                         WHERE catid='.$pageId);
+                                    while(!$objResult->EOF) {
+                                        $objDatabase->Execute('
+                                            DELETE FROM '.DBPREFIX.'content_logfile
+                                                WHERE   history_id ='.$objResult->fields["id"]);
+                                        $objResult->MoveNext();
+                                    }
+                                }
                             $objDatabase->Execute('
                                 UPDATE '.DBPREFIX.'content_navigation_history
                                    SET changelog='.time().'
@@ -2483,7 +2500,6 @@ class ContentManager
                         }
 
                         $boolDelete = true;
-                        $objFWUser = FWUser::getFWUserObject();
                         if ($this->boolHistoryEnabled) {
                             /*
                              * Only Admin and Page Owner can delete a page
@@ -2756,10 +2772,14 @@ class ContentManager
         if ($langId == 0) {
             $langId = $this->firstActiveLang;
         }
+        /*
+         * Display only validated pages in 'Ãœbergeordnete Seite' select menu
+         */
         $objResult = $objDatabase->Execute("
             SELECT catid, parcat, catname, backend_access_id
               FROM ".DBPREFIX."content_navigation
              WHERE `lang`=".$langId."
+               AND is_validated = '1'
              ORDER BY parcat ASC, displayorder ASC");
         if (!$objResult)
             return "content::navigation() database error";
